change: test files + fixes
