+++ 3Dataviz/frontend/__test__/components/ut/AveragePlane.test.tsx
@@ -2,8 +2,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import ReactThreeTestRenderer from "@react-three/test-renderer";
 import React from "react";
 import * as THREE from "three";
-import { AveragePlaneProps } from "../../../src/components/Planes/props/AveragePlaneProps";
-import AveragePlane from "../../../src/components/Planes/AveragePlane";
+import { averagePlaneProps } from "../../../src/components/planes/props/averagePlaneProps";
+import AveragePlane from "../../../src/components/planes/averagePlane";
 
 vi.mock("react-redux", () => ({
   ...vi.importActual("react-redux"),
@@ -16,8 +16,7 @@ describe("AveragePlane", () => {
     vi.clearAllMocks();
   });
   it("Renderizza average plane senza crashare", async () => {
-    const mockProp: AveragePlaneProps = {
-      isVisible: true,
+    const mockProp: averagePlaneProps = {
       position: new THREE.Vector3(0, 0, 0),
       size: [1, 1],
     };
