+++ 3Dataviz/frontend/__test__/components/ut/Options.test.tsx
@@ -1,12 +1,12 @@
 import { describe, expect, it, vi } from "vitest";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
-import Options from "../../../src/components/UI/Options/Options";
+import Options from "../../../src/components/UI/options/options";
 import React from "react";
 import { Provider } from "react-redux";
 import configureMockStore from "redux-mock-store";
 import { thunk } from "redux-thunk";
-import { AppDispatch, RootState } from "../../../src/app/Store";
-import { CreateMockRootState } from "../../utils/StateMockCreator";
+import { AppDispatch, RootState } from "../../../src/app/store";
+import { createMockRootState } from "../../utils/stateMockCreator";
 import gsap from "gsap";
 
 const middlewares = [thunk];
@@ -15,7 +15,7 @@ const mockStore = configureMockStore<RootState, AppDispatch>(middlewares);
 describe("Options", () => {
   it("le opzioni devono aprirsi quando si clicca sul bottone apposito", async () => {
     render(
-      <Provider store={mockStore(CreateMockRootState())}>
+      <Provider store={mockStore(createMockRootState())}>
         <Options />
       </Provider>,
     );
@@ -36,7 +36,7 @@ describe("Options", () => {
 
   it("La tabella deve chiudersi quando si clicca sul bottone apposito", async () => {
     render(
-      <Provider store={mockStore(CreateMockRootState())}>
+      <Provider store={mockStore(createMockRootState())}>
         <Options />
       </Provider>,
     );
@@ -59,7 +59,7 @@ describe("Options", () => {
   describe("Filter", () => {
     it("la Ui per il filtraggio rispetto al valor medio globale deve essere visibile al caricamento del dataset", async () => {
       render(
-        <Provider store={mockStore(CreateMockRootState())}>
+        <Provider store={mockStore(createMockRootState())}>
           <Options />
         </Provider>,
       );
@@ -69,7 +69,7 @@ describe("Options", () => {
 
     it("la Ui per il filtraggio rispetto al valor medio globale deve essere visibile al caricamento del dataset", async () => {
       render(
-        <Provider store={mockStore(CreateMockRootState())}>
+        <Provider store={mockStore(createMockRootState())}>
           <Options />
         </Provider>,
       );
@@ -80,7 +80,7 @@ describe("Options", () => {
     it("il bottone deve eseguire il filtraggio rispetto al valor medio (inferiori)", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -106,7 +106,7 @@ describe("Options", () => {
     it("il bottone deve eseguire il filtraggio rispetto al valor medio (superiori)", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -137,7 +137,7 @@ describe("Options", () => {
   describe("NFilter", () => {
     it("la Ui per il filtraggio per i top e bottom N deve essere visibile al caricamento del dataset", async () => {
       render(
-        <Provider store={mockStore(CreateMockRootState())}>
+        <Provider store={mockStore(createMockRootState())}>
           <Options />
         </Provider>,
       );
@@ -147,7 +147,7 @@ describe("Options", () => {
     it("il bottone deve eseguire il filtraggio dei bottom N valori", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -174,7 +174,7 @@ describe("Options", () => {
     it("il bottone deve eseguire il filtraggio dei top N valori", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -204,7 +204,7 @@ describe("Options", () => {
     it("il bottone deve eseguire il filtraggio dei top o bottom 0 valori se il valore inserito non e` un numero o e` <0", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -237,7 +237,7 @@ describe("Options", () => {
   describe("FilterModOptions", () => {
     it("la Ui per cambiare le opzioni di filtraggio siano visibili al caricamento del dataset", async () => {
       render(
-        <Provider store={mockStore(CreateMockRootState())}>
+        <Provider store={mockStore(createMockRootState())}>
           <Options />
         </Provider>,
       );
@@ -247,7 +247,7 @@ describe("Options", () => {
     it("quando cambio l`algoritmo di filtraggio in superiore il componente deve aggiornare lo stato con la relativa action", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -273,7 +273,7 @@ describe("Options", () => {
     it("quando cambio l`algoritmo di filtraggio in inferiori il componente deve aggiornare lo stato con la relativa action", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -303,7 +303,7 @@ describe("Options", () => {
   describe("AveragePlaneOption", () => {
     it("la Ui per cambiare le opzioni di visualizzazione del piano medio siano visibili al caricamento del dataset", async () => {
       render(
-        <Provider store={mockStore(CreateMockRootState())}>
+        <Provider store={mockStore(createMockRootState())}>
           <Options />
         </Provider>,
       );
@@ -314,7 +314,7 @@ describe("Options", () => {
     it("quando cambio la visibilita` in 'abilitato' del piano medio globale il componente richiami la action corretta per aggiornare lo stato", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -341,7 +341,7 @@ describe("Options", () => {
     it("quando cambio la visibilita` in 'disabilitato' del piano medio globale il componente richiami la action corretta per aggiornare lo stato", async () => {
       const dispatch = vi.fn();
       const store = mockStore(
-        CreateMockRootState({
+        createMockRootState({
           data: {
             data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
             average: 1,
@@ -368,7 +368,7 @@ describe("Options", () => {
 
   it("la Ui per resettare il filtraggio sia visibile al caricamento del dataset", async () => {
     render(
-      <Provider store={mockStore(CreateMockRootState())}>
+      <Provider store={mockStore(createMockRootState())}>
         <Options />
       </Provider>,
     );
@@ -379,7 +379,7 @@ describe("Options", () => {
   it("il bottone di reset deve eseguire un`azione di reset", async () => {
     const dispatch = vi.fn();
     const store = mockStore(
-      CreateMockRootState({
+      createMockRootState({
         data: {
           data: [{ id: 1, show: true, x: 0, y: 1, z: 0 }],
           average: 1,
